Moving debugging code to sandbox where it probably belongs
